refactor moved sub folders in review to component folder in practitioner
