+++ src/app/modules/practitioner-registration/pages/facility-info/facility-info.component.html
@@ -1,4 +1,3 @@
 <h1>{{pageTitle}}</h1>
-<hr>
 
 <bcp-core-facility-info name='facInfo' [formGroup]='formGroup'></bcp-core-facility-info>
