+++ src/version.GENERATED.ts
@@ -1,9 +1,9 @@
 // DO NOT DELETE OR APP WILL FAIL TO COMPILE! Generated from version.js
-export const gitCommit = '0aa9a18';
-export const buildTime = '12/11/2019 at 3:06:24 PM';
+export const gitCommit = 'ed0a5be';
+export const buildTime = '12/11/2019 at 6:18:52 PM';
 /** App version retrieved from package.json. */
 export const projectVersion = '0.9.0';
 /** Human readable message  */
-export const message = 'Canonical Version: 0aa9a18 - 12/11/2019 at 3:06:24 PM (v0.9.0) \nhttps://github.com/bcgov/moh-bcp/commits/0aa9a18';
+export const message = 'Canonical Version: ed0a5be - 12/11/2019 at 6:18:52 PM (v0.9.0) \nhttps://github.com/bcgov/moh-bcp/commits/ed0a5be';
 /** If true, other values should be present. */
 export const success = true;
